If the user issues a TEARDOWN request, we now stop sending them any video frames until the next issue a PLAY request.
This is done to help avoid race conditions in Obj-C land, with the RTSP Client sometimes attempting to being destroyed on one of its own threads. This happens when the last strong SurgeRtspPlayer pointer is removed while the player is currently rendering a frame - which at some point uses a weak Player reference, causing a weak retain and ARC ends up running on the wrong thread. A better solution will be needed in the future.