Merged in mm_realmRegex (pull request #12)
Expanding regex for parsing realm from Www-Authenticate header

Expanding the regex for the “realm” portion of the Www-Authenticate header.  Some cameras include various characters not captured by the previous version of the regex, which would result in the regex discarding a portion of the realm value, subsequently messing up MD5 hashing for authentication.

Approved-by: Paul Shields <paul.shields@instil.co>